requirements.txt
Flask
DeepFace
librosa
transformers
supabase
Pillow
numpy
scipy
torch
requests 